+++ src/core/src/main/java/org/sakuli/actions/TestCaseAction.java
@@ -63,8 +63,10 @@ public class TestCaseAction {
      * Set the warning and critical Time to the specific test case.
      *
      * @param testCaseID   current ID of the test case
-     * @param warningTime  warning threshold in seconds
-     * @param criticalTime critical threshold in seconds
+     * @param warningTime  warning threshold in seconds. If the threshold is set to 0,
+     *                     the execution time will never exceed, so the state will be always OK!
+     * @param criticalTime critical threshold in seconds. If the threshold is set to 0,
+     *                     the execution time will never exceed, so the state will be always OK!
      * @param imagePaths   multiple paths to images
      */
     @LogToResult(message = "init a new test case")
@@ -77,8 +79,10 @@ public class TestCaseAction {
      * Set the warning and critical Time to the specific test case.
      *
      * @param testCaseID   current ID of the test case
-     * @param warningTime  warning threshold in seconds
-     * @param criticalTime critical threshold in seconds
+     * @param warningTime  warning threshold in seconds. If the threshold is set to 0,
+     *                     the execution time will never exceed, so the state will be always OK!
+     * @param criticalTime critical threshold in seconds. If the threshold is set to 0,
+     *                     the execution time will never exceed, so the state will be always OK!
      * @param imagePaths   multiple paths to images
      */
     @LogToResult(message = "init a new test case")
@@ -180,7 +184,7 @@ public class TestCaseAction {
      * @param stepName    name of this step
      * @param startTime   start time in milliseconds
      * @param stopTime    end time in milliseconds
-     * @param warningTime warning threshold in seconds
+     * @param warningTime warning threshold in seconds. If the threshold is set to 0, the execution time will never exceed, so the state will be always OK!
      * @throws SakuliException
      */
     @LogToResult(message = "add a step to the current test case")